+++ Front/src/components/Menu.js
@@ -1,0 +1,5 @@
+const Menu = () => {
+
+}
+
+export default Menu;
