+++ Front/src/pages/Main.js
@@ -1,0 +1,21 @@
+import { View, Text, StatusBar, StyleSheet } from "react-native";
+
+const Main = () => {
+  return (
+    <View style={styles.container}>
+      <Text>Open up App.js to start working on your app!</Text>
+      <StatusBar style="auto" />
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+      backgroundColor: '#fff',
+      alignItems: 'center',
+      justifyContent: 'center',
+    },
+  });
+
+export default Main;
